Added parser2 tests

+++ source/ast/type_reference.hpp
@@ -14,7 +14,7 @@ namespace cmsl
             cmsl::string_view to_string() const
             {
                 const auto b = begin.str().data();
-                return cmsl::string_view{ b, end.str().cend() - b };
+                return cmsl::string_view{ b, static_cast<size_t>(end.str().cend() - b)};
             }
         };
     }
